fix prob handling

+++ elements/empower/empowerlvapmanager.cc
@@ -656,7 +656,7 @@ void EmpowerLVAPManager::send_lvap_stats_response(EtherAddress lvap, uint32_t lv
 		assert (ptr <= end);
 		lvap_stats_entry *entry = (lvap_stats_entry *) ptr;
 		entry->set_rate(nfo->rates[i]);
-		entry->set_prob(nfo->probability[i]);
+		entry->set_prob((uint32_t) nfo->probability[i]);
 		ptr += sizeof(struct lvap_stats_entry);
 	}
